pyproject.toml
[tool.poetry]
name = "akinon-case"
version = "0.1.0"
description = ""
authors = ["Yankı Sırtaş <yankisirtas@hotmail.com>"]
readme = "README.md"
packages = [{include = "akinon_case"}]

[tool.poetry.dependencies]
python = "^3.11"
djangorestframework = "^3.14.0"
markdown = "^3.4.4"
django-filter = "^23.2"
black = "^23.7.0"
flake8 = "^6.1.0"
requests = "^2.31.0"
python-decouple = "^3.8"
celery = "^5.3.1"
django-celery-beat = "^2.5.0"
django-celery-results = "^2.5.1"
redis = "^4.6.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
